little refactoring, changed the functions to dynamic_optimizer file, added doc
