+++ avalanche/models/dynamic_optimizers.py
@@ -17,6 +17,13 @@
 from collections import defaultdict
 
 
+def compare_keys(old_dict, new_dict):
+    not_in_new = list(set(old_dict.keys()) - set(new_dict.keys()))
+    in_both = list(set(old_dict.keys()) & set(new_dict.keys()))
+    not_in_old = list(set(new_dict.keys()) - set(old_dict.keys()))
+    return not_in_new, in_both, not_in_old
+
+
 def reset_optimizer(optimizer, model):
     """Reset the optimizer to update the list of learnable parameters.
 
@@ -29,39 +36,102 @@ def reset_optimizer(optimizer, model):
     """
     assert len(optimizer.param_groups) == 1
     optimizer.state = defaultdict(dict)
-    optimizer.param_groups[0]["params"] = list(model.parameters())
+
+    parameters = []
+    optimized_param_id = {}
+    for n, p in model.named_parameters():
+        if p.requires_grad:
+            optimized_param_id[n] = p
+            parameters.append(p)
+
+    optimizer.param_groups[0]["params"] = parameters
+
+    return optimized_param_id
 
 
-def update_optimizer(optimizer, old_params, new_params, reset_state=True):
-    """Update the optimizer by substituting old_params with new_params.
+def update_optimizer(
+        optimizer, 
+        new_params, 
+        optimized_params, 
+        reset_state=False
+):
+    """Update the optimizer by adding new parameters,
+    removing removed parameters, and adding new parameters
+    to the optimizer, for instance after model has been adapted
+    to a new task. The state of the optimizer can also be reset,
+    it will be reset for the modified parameters.
 
-    :param old_params: List of old trainable parameters.
-    :param new_params: List of new trainable parameters.
-    :param reset_state: Wheter to reset the optimizer's state.
-        Defaults to True.
-    :return:
+    Newly added parameters are added by default to parameter group 0
+
+    :param new_params: Dict (name, param) of new parameters
+    :param optimized_params: Dict (name, param) of 
+        currently optimized parameters (returned by reset_optimizer)
+    :param reset_state: Wheter to reset the optimizer's state (i.e momentum).
+        Defaults to False.
+    :return: Dict (name, param) of optimized parameters
     """
-    for old_p, new_p in zip(old_params, new_params):
+    not_in_new, in_both, not_in_old = compare_keys(optimized_params, new_params)
+    keys_to_remove = []
+    # Change reference to already existing parameters
+    # i.e growing IncrementalClassifier
+    for key in in_both:
+        old_p_hash = optimized_params[key]
+        new_p = new_params[key]
+        # Look for old parameter id in current optimizer
         found = False
-        # iterate over group and params for each group.
         for group in optimizer.param_groups:
             for i, curr_p in enumerate(group["params"]):
-                if hash(curr_p) == hash(old_p):
-                    # update parameter reference
-                    group["params"][i] = new_p
+                if id(curr_p) == id(old_p_hash):
                     found = True
+
+                    if id(curr_p) != id(new_p):
+                        group["params"][i] = new_p
+                        optimized_params[key] = new_p
+                        if not new_p.requires_grad:
+                            keys_to_remove.append(key)
+                        if curr_p in optimizer.state:
+                            optimizer.state[new_p] = optimizer.state.pop(curr_p)
                     break
-            if found:
-                break
         if not found:
-            raise Exception(
-                f"Parameter {old_params} not found in the "
-                f"current optimizer."
-            )
+            raise Exception(f"Parameter {key} expected but "
+                            "not found in the optimizer")
+
+    # Remove parameters that are not here anymore
+    # This should not happend in most use case
+    for key in not_in_new:
+        old_p_hash = optimized_params[key]
+        found = False
+        for i, group in enumerate(optimizer.param_groups):
+            keys_to_remove.append([])
+            for j, curr_p in enumerate(group["params"]):
+                if id(curr_p) == id(old_p_hash):
+                    found = True
+                    keys_to_remove[i].append((j, curr_p))
+                    optimized_params.pop(key)
+                    break
+        if not found:
+            raise Exception(f"Parameter {key} expected but " 
+                            "not found in the optimizer")
+
+    for i, idx_list in enumerate(keys_to_remove):
+        for (j, p) in sorted(idx_list, key=lambda x: x[0], reverse=True):
+            del optimizer.param_groups[i]["params"][j]
+            if p in optimizer.state:
+                optimizer.state.pop(p)
+
+    # Add newly added parameters (i.e Multitask, PNN)
+    # by default, add to param groups 0
+    for key in not_in_old:
+        new_p = new_params[key]
+        if new_p.requires_grad:
+            optimizer.param_groups[0]["params"].append(new_p)
+            optimized_params[key] = new_p
+            optimizer.state[new_p] = {}
+
     if reset_state:
-        # State contains parameter-specific information.
-        # We reset it because the model is (probably) changed.
         optimizer.state = defaultdict(dict)
+
+    return optimized_params
 
 
 def add_new_params_to_optimizer(optimizer, new_params):
@@ -70,3 +140,6 @@ def add_new_params_to_optimizer(optimizer, new_params):
     :param new_params: list of trainable parameters
     """
     optimizer.add_param_group({"params": new_params})
+
+
+__all__ = ["add_new_params_to_optimizer", "reset_optimizer", "update_optimizer"]
